+++ src/components/Wallet/Deposit/FailedDeposits.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { View, Text, TouchableOpacity, TextInput } from 'react-native';
 import { ArrowRight, Camera as CameraIcon } from 'react-native-feather';
+import bip21 from 'bip21';
 
 import Spinner from '../../reusable/Spinner';
 import { breezGetFailedDeposits, breezRefundDeposit } from '../../../api/breez';
@@ -10,6 +11,7 @@ import FailedDepositItem from './FailedDepositItem';
 import Button from '../../reusable/Button';
 import FeePicker from '../../shared/FeePicker';
 import QrScanner from '../../reusable/QrScanner';
+import { refundDeposit } from '../../../database/database';
 
 interface Props {
   setError: (error: string) => void;
@@ -29,9 +31,13 @@ const FailedDeposits = ({ setError }: Props) => {
   };
 
   const onQrScanned = (value: string) => {
-    if (isBtcAddress(value)) {
-      setRefundAddress(value);
+    // validate address and set value
+    try {
+      const decoded = bip21.decode(value);
+      setRefundAddress(decoded.address);
       setActiveCamera(false);
+    } catch (_) {
+      console.error('found invalid BIP21');
     }
   };
 
@@ -50,12 +56,18 @@ const FailedDeposits = ({ setError }: Props) => {
 
     for (const swap of failedDeposits) {
       breezRefundDeposit(refundAddress, swap, fee).then(() => {
-        setFormError(undefined);
-        setFailedDeposits(
-          failedDeposits?.filter(s => s.paymentHash !== swap.paymentHash),
-        );
+        refundDeposit(swap)
+          .then(() => {
+            console.log('Refunded deposit', swap);
+          })
+          .catch(() => {
+            console.error('Failed write refunded state to deposit', swap);
+          });
       });
     }
+
+    setFailedDeposits([]);
+    setFormError(undefined);
   };
 
   React.useEffect(() => {
@@ -88,47 +100,61 @@ const FailedDeposits = ({ setError }: Props) => {
     return null;
   }
 
+  const buttonDisabled = !isBtcAddress(refundAddress) || !fee || loading;
+
   return (
-    <View className="flex flex-col mx-auto w-full py-4 px-2 pb-16">
+    <View className="flex flex-col w-full py-4 px-2 pb-16">
       <QrScanner
         onClose={() => setActiveCamera(false)}
         visible={activeCamera}
         onQrCodeScanned={onQrScanned}
       />
-      <Text className="text-xl">Depositi falliti</Text>
+      <Text className="text-center text-2xl text-brandAlt">
+        Depositi Falliti
+      </Text>
       <View>
         {failedDeposits.map((deposit, i) => (
           <FailedDepositItem key={i} swap={deposit} />
         ))}
       </View>
-      <Text className="text-lg">
+      <Text className="text-lg text-text">
         Imposta un indirizzo BTC sul quale effettuare il rimborso
       </Text>
-      <View className="mt-4 pr-4 flex flex-row items-center justify-center bg-gray-50 border border-gray-300 h-min">
-        <TextInput
-          className="text-text text-sm rounded-lg focus:ring-brand focus:border-brand p-4 focus-visible:outline-none w-4/6"
-          placeholder="Indirizzo BTC"
-          onChangeText={setRefundAddress}
-          defaultValue={refundAddress}
-        />
-        <TouchableOpacity onPress={onScanQrCode}>
-          <CameraIcon className="w-8 h-8 text-brandAlt" />
-        </TouchableOpacity>
-      </View>
-      <View className="mt-4 pr-20 flex flex-row items-center justify-center bg-gray-50 border border-gray-300 h-min">
-        <FeePicker
-          className="w-full"
-          fee={fee}
-          onFeeChanged={setFee}
-          onError={setError}
-        />
-      </View>
-      <View className="flex flex-col justify-center items-center">
-        {formError && <Text className="text-red-500">{formError}</Text>}
-        <Button.Primary onPress={onRefund} disabled={loading}>
-          <Text className="text-white text-2xl">Rimborsa depositi</Text>
-          <ArrowRight className=" text-white" width={24} height={24} />
-        </Button.Primary>
+      <View className="w-page mx-auto">
+        <View className="mt-4 w-full">
+          <Text className=" text-text py-1 text-lg">Indirizzo BTC</Text>
+          <View className="px-4 flex flex-row items-center justify-center bg-gray-50 border border-gray-300 h-min">
+            <TextInput
+              className="text-text text-sm rounded-lg focus:ring-brand focus:border-brand p-4 focus-visible:outline-none w-full"
+              placeholder="bc1..."
+              onChangeText={setRefundAddress}
+              defaultValue={refundAddress}
+            />
+            <TouchableOpacity onPress={onScanQrCode}>
+              <CameraIcon className="w-8 h-8 text-brandAlt" />
+            </TouchableOpacity>
+          </View>
+        </View>
+        <View className="mt-4 w-full">
+          <Text className=" text-text py-1 text-lg">Seleziona fee</Text>
+          <View className="flex flex-row items-center justify-center h-[55px] w-full">
+            <View className="left-0 right-0 mx-auto border-gray-300 border bg-gray-50 h-full w-full">
+              <FeePicker
+                className="text-text w-full"
+                fee={fee}
+                onFeeChanged={setFee}
+                onError={setError}
+              />
+            </View>
+          </View>
+        </View>
+        <View className="flex flex-col justify-center items-center">
+          {formError && <Text className="text-red-500">{formError}</Text>}
+          <Button.Primary onPress={onRefund} disabled={buttonDisabled}>
+            <Text className="text-white text-2xl">Rimborsa depositi</Text>
+            <ArrowRight className=" text-white" width={24} height={24} />
+          </Button.Primary>
+        </View>
       </View>
     </View>
   );