+++ CHANGELOG.md
@@ -1,9 +1,21 @@
 # Changelog
 
 - [Changelog](#changelog)
+  - [0.2.1](#021)
   - [0.2.0](#020)
   - [0.1.2](#012)
   - [0.1.1](#011)
+
+## 0.2.1
+
+Released on 13/12/2023
+
+- [Issue 25](https://github.com/SatoshisPay/satoshispay/issues/25): Deposits are now logged in the deposit form.
+  - Added amount for deposits (BIP21).
+- [Issue 28](https://github.com/SatoshisPay/satoshispay/issues/28): Check for minimum amount to withdraw before sending withdraw
+- [Issue 32](https://github.com/SatoshisPay/satoshispay/issues/32): the transactions CSV couldn't be located.
+- [Issue 35](https://github.com/SatoshisPay/satoshispay/issues/35): BIP21 QR code support
+- [Issue 41](https://github.com/SatoshisPay/satoshispay/issues/41): It is now possible to buy Bitcoin in-app through the Breez providers (Moonpay atm)
 
 ## 0.2.0